+++ src/app/core/services/api/file/image.service.ts
