+++ src/app/core/services/api/rights/rights.service.ts
