+++ src/app/core/services/auth.service.ts
@@ -8,7 +8,7 @@ import { AuthenticationResponse } from '@data/api/auth-service/models/authentica
 import { AuthApiService } from '@data/api/auth-service/services/auth-api.service';
 import { CredentialsApiService } from '@data/api/user-service/services/credentials-api.service';
 import { CreateCredentialsRequest } from '@data/api/user-service/models/create-credentials-request';
-import { UserService } from '@app/services/user.service';
+import { UserService } from '@app/services/api/user/user.service';
 import { HttpErrorResponse } from '@angular/common/http';
 import { OperationResultResponseCredentialsResponse } from '@data/api/user-service/models/operation-result-response-credentials-response';
 import { User } from '@app/models/user/user.model';
