+++ src/app/modules/admin/components/department-card/department-card.component.ts
@@ -3,7 +3,7 @@ import { Sort } from '@angular/material/sort';
 import { NetService } from '@app/services/net.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
-import { UserService } from '@app/services/user.service';
+import { UserService } from '@app/services/api/user/user.service';
 import { UserInfo } from '@data/api/user-service/models/user-info';
 import { PageEvent } from '@angular/material/paginator';
 import { DepartmentInfo } from '@data/api/company-service/models/department-info';
