+++ src/app/modules/auth/components/forgot-password/forgot-password.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 
-import { UserService } from '@app/services/user.service';
+import { UserService } from '@app/services/api/user/user.service';
 
 @Component({
 	selector: 'do-forgot-password',
