+++ src/app/modules/user/components/attendance/attendance.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 
 import { DatePeriod } from '@data/models/date-period';
 import { IDailyHoursData } from '../gradient-graphics/gradient-graphics.component';
-import { ProjectService } from '@app/services/project.service';
+import { ProjectService } from '@app/services/api/project/project.service';
 import { Observable, of } from 'rxjs';
 import { Project } from '@app/models/project/project.model';
 import { map } from 'rxjs/operators';
