+++ src/app/shared/component/content-container/content-container.component.ts
@@ -1,6 +1,6 @@
 import { Component, ElementRef, OnInit, ViewChild } from '@angular/core';
 
-import { UserService } from '@app/services/user.service';
+import { UserService } from '@app/services/api/user/user.service';
 import { AuthService } from '@app/services/auth.service';
 import { Router } from '@angular/router';
 import { User } from '@app/models/user/user.model';
